requirements.txt
langchain
pinecone-client 
python-dotenv
openai
tiktoken
streamlit
python-docx
azure-core